code/requirements.txt
numpy~=1.26.4
pandas~=2.2.2
transformers~=4.41.2
tqdm~=4.66.4
torch~=2.3.1
torchvision~=0.18.1
Pillow~=10.4.0
scikit-learn~=1.5.1
scipy~=1.10.1
datasets~=2.19.2